backend/requirements.txt
flask
flask-cors
langchain
faiss-cpu
openai
PyPDF2
python-docx